feat: Add Hackathon technology to MainTechFilter component
This commit adds the Hackathon technology to the MainTechFilter component in order to display it as an option for filtering mini projects. The Hackathon technology is represented by the AiFillTrophy icon and has a yellow color theme. This change improves the user experience by providing a new filtering option for Hackathon projects.